changed types from 'int' to 'long'

+++ Search/Binary/src/main/java/ru/mrchebik/BinarySearch.java
@@ -4,13 +4,13 @@ package ru.mrchebik;
  * Created by mrchebik on 13.03.17.
  */
 public class BinarySearch {
-    private static final int[] array = new int[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20};
+    private static final long[] array = new long[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20};
 
     public static void main(String[] args) {
         System.out.println(binary(array, 9));
     }
 
-    public static int binary(int[] array, int searchKey) {
+    public static int binary(long[] array, int searchKey) {
         int lowerBound = 0;
         int upperBound = array.length - 1;
         int current;
